Move DefaultOutputSnapshotter to the right package
